dev: test workspace with memory storage

+++ src/ghoshell_common/contracts/workspace.py
@@ -2,7 +2,7 @@ import os
 from abc import ABC, abstractmethod
 from typing import Optional
 
-from ghoshell_common.contracts.storage import FileStorage, DefaultFileStorage
+from ghoshell_common.contracts.storage import Storage, DefaultFileStorage, MemoryStorage
 from ghoshell_container import Container, Provider
 from os.path import abspath
 import shutil
@@ -17,28 +17,28 @@ class Workspace(ABC):
     """
 
     @abstractmethod
-    def root(self) -> FileStorage:
+    def root(self) -> Storage:
         """
         workspace 根 storage.
         """
         pass
 
     @abstractmethod
-    def configs(self) -> FileStorage:
+    def configs(self) -> Storage:
         """
         配置文件存储路径.
         """
         pass
 
     @abstractmethod
-    def runtime(self) -> FileStorage:
+    def runtime(self) -> Storage:
         """
         运行时数据存储路径.
         """
         pass
 
     @abstractmethod
-    def assets(self) -> FileStorage:
+    def assets(self) -> Storage:
         """
         数据资产存储路径.
         """
@@ -50,18 +50,21 @@ class LocalWorkspace(Workspace):
     def __init__(self, workspace_dir: str):
         workspace_dir = abspath(workspace_dir)
         self._ws_root_dir = workspace_dir
-        self._root = DefaultFileStorage(workspace_dir)
+        if not self._ws_root_dir:
+            self._root = MemoryStorage(self._ws_root_dir)
+        else:
+            self._root = DefaultFileStorage(workspace_dir)
 
-    def root(self) -> FileStorage:
+    def root(self) -> Storage:
         return self._root
 
-    def configs(self) -> FileStorage:
+    def configs(self) -> Storage:
         return self._root.sub_storage("configs")
 
-    def runtime(self) -> FileStorage:
+    def runtime(self) -> Storage:
         return self._root.sub_storage("runtime")
 
-    def assets(self) -> FileStorage:
+    def assets(self) -> Storage:
         return self._root.sub_storage("assets")
 
 
@@ -74,15 +77,15 @@ class LocalWorkspaceProvider(Provider[Workspace]):
     ):
         if workspace_dir == "":
             # 使用脚本运行的路径作为 workspace.
-            workspace_dir = os.path.join(abspath(os.getcwd()), ".ghoshell_ws/")
-        self._ws_dir = abspath(workspace_dir)
+            workspace_dir = abspath(workspace_dir)
+        self._ws_dir = workspace_dir
         self._stub_dir = abspath(stub_dir) if stub_dir else None
 
     def singleton(self) -> bool:
         return True
 
     def factory(self, con: Container) -> Optional[Workspace]:
-        if self._stub_dir and not os.path.exists(self._stub_dir):
+        if self._ws_dir and self._stub_dir and not os.path.exists(self._stub_dir):
             os.makedirs(self._stub_dir)
             shutil.copytree(self._stub_dir, self._ws_dir)
         return LocalWorkspace(self._ws_dir)
